Patched Crystal version 0.1 beta
~~0.1~~
Ghost type attacks take advantage of Special Attack stat instead of
Attack stat. (Im not sure if this counts as a bug but for me it seems
weird that Ghost type attacks were physical instead of special, so i
changed it.)
Haunter now evolves into Gengar at level 40. (This was done for testing
purposes and kept as is)
Moon ball and Fast ball work as intended now.
Status ailments increase a pokemon's catch rate as they should.