Pokemon Patched Crystal - Version 3.0
This is yet another "big" update.

This version uses the latest repo of pokecrystal (as of 14/02/2022) along with a few tweaks and a fix on the surfing crashing bug, as always, keep a backup of your save just in case but it should work fine, more info on the readme.