Pokemon Patched Crystal - Version 3.8
Smashing rocks now has a chance to give the player one of the following items:

-Nugget
-Star Piece
-Big Pearl
-Pearl
-Brick Piece
-Metal Coat
-Hard Stone
-Soft Sand
-Thick Club
-Dome Fossil
-Helix Fossil
-Old Amber

With the existance of fossils from the above addition, you can no longer find Kabuto and Omanyte in the wild.

You can revive your fossils with the help of the new Scientist NPC inside the Ruins of Alph Research Center.

The time of day (Day or Night) can now be seen with an icon on the main battle menu.

When picking up or receiving TMs as gifts the name of the move will also be shown.

The Mew you are given after defeating Cooltrainer Rose now holds Stardust.